requirements.txt
win32com.client 
speech_recognition 
os
AppOpener 
groq 
datetime
webbrowser